requirements.txt
gunicorn==20.1.0
Django==4.1.4
django-filter==22.1
django-redis==5.2.0
djangorestframework==3.14.0
Markdown==3.3.4